fix surrogate batching
