+++ brainpy/_src/math/surrogate/__init__.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 
 
-from .base import *
 from ._one_input_new import *
 from ._two_inputs import *
